+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release Notes for Easy Gallery
 
+## 1.0.2
+- Added the getAllAssets service
+- Added the getAllFolders service
+- Added the getAllObjects service
+- Added the getPaths service
+
 ## 1.0.1
 - Changed filter behaviour
 - Readme.md updated